librarian-mcp-server/requirements.txt
python-dotenv
langchain>=0.2.0
langchain-core
langchain-community
ddgs
duckduckgo-search
langchain-google-genai
langchain-pinecone
pinecone-client>=4.0.0
beautifulsoup4
requests
langchain-text-splitters
tavily-python
crawl4ai
fastapi
uvicorn
asyncio
certifi
fastembed
langchain-groq
python-multipart
nest-asyncio